simplify directory structure to single python script
