+++ diatomicmolecule.py
@@ -1,4 +1,4 @@
-"""Function for building a diatomic molecule."""
+import json
 
 
 def create_diatomic_molecule_geometry(species1, species2, bond_length):
@@ -21,3 +21,10 @@ def create_diatomic_molecule_geometry(species1, species2, bond_length):
     }
 
     return geometry
+
+
+def create_diatomic_molecule(species1, species2, bond_length):
+    geometry = create_diatomic_molecule_geometry(species1, species2, bond_length)
+    geometry["schema"] = "molecular_geometry"
+    with open("molecule.json", "w") as f:
+        f.write(json.dumps(geometry))
